+++ app/screens/detailCategory/components/resumedTab/index.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { ScrollView } from "react-native";
+import { FlatList } from "react-native";
 import { Results } from "realm/dist/bundle";
 
 import Balance from "../../../../models/schemas/BalanceSchema";
@@ -25,9 +25,11 @@ const ResumedTabComponent: React.FC<ResumedTabComponentProps> = ({ filteredBalan
   return (
     <>
       <SectionHeader title='Total' value={convertToMoney(allExpensesResult)}/>
-      <ScrollView showsVerticalScrollIndicator={false}>
-        { Array.from(filteredBalanceList).reverse().map((balance) => renderItem(balance)) }
-      </ScrollView>
+      <FlatList 
+        showsVerticalScrollIndicator={false} 
+        data={Array.from(filteredBalanceList).reverse()} 
+        renderItem={({ item }) => renderItem(item)}
+      />
     </>
   )
 }